streamlit-ui/pyproject.toml
[project]
name = "convergefi-streamlit-ui"
version = "0.1.0"
description = "ConvergeFi Streamlit testing interface"
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "requests",
    "pillow",
    "pandas",
    "plotly",
    "python-dotenv",
    "pydantic-settings>=2.11.0",
    "httpx>=0.28.1",
]
